requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
playwright==1.47.0
python-dotenv==1.0.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0